Move land shared types and artifact-mode landing out of land/mod.rs
land/types.rs owns the plan/run data model (LandPlan, LandRun, step
kinds/status, step outcomes); land/artifact.rs owns the no-workspace
artifact landing path. land/mod.rs keeps the entry points, path
resolution, and provider/PR helpers shared by check/execute/plan.
Pure moves, no behavior change.

Knit-Group: kg_20260612_e5364b
Knit-Bundle: open-source-prep
