+++ src/commands/land/mod.rs
@@ -11,309 +11,33 @@
 //! - [`update`] implements `knit land update` (merge base into feature branches)
 //! - [`display`] renders plans, run status, and PR/check state
 
+mod artifact;
 mod check;
 mod display;
 mod execute;
 mod plan;
+mod types;
 mod update;
 mod validate;
 
+pub use artifact::apply_land_from_artifact;
 pub use check::check_landing;
 pub(crate) use check::{assess_landing_readiness, print_readiness_row};
 
 use crate::advice;
 use crate::checkout::{checkout_dir, ensure_expected_branch};
-use crate::ids::node_id;
-use crate::model::{BundleNode, DeployCheckoutUpdate, DeployMode, MergeMethod, RepoEntry};
+use crate::model::RepoEntry;
 use crate::output as out;
-use crate::providers::{self, publication_for_repo, Forge, PrTarget, PullRequest};
+use crate::providers::{self, Forge, PrTarget, PullRequest};
 use crate::store::{
     load_active_bundle, load_active_bundle_for_update, read_json, write_json, ActiveBundle,
 };
 use crate::time::now_iso;
 use anyhow::{bail, Context, Result};
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 
-const LAND_PLAN_KIND: &str = "KnitLandPlan";
-const LAND_RUN_KIND: &str = "KnitLandRun";
-const DEFAULT_LAND_PROVIDER: &str = "github";
-
-/// Kind of a land plan step. Serialized snake_case to match the editable plan
-/// files (`merge_pr`, `wait_checks`, ...).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum LandStepKind {
-    MergePr,
-    WaitChecks,
-    Run,
-    Deploy,
-}
-
-impl LandStepKind {
-    fn as_str(self) -> &'static str {
-        match self {
-            LandStepKind::MergePr => "merge_pr",
-            LandStepKind::WaitChecks => "wait_checks",
-            LandStepKind::Run => "run",
-            LandStepKind::Deploy => "deploy",
-        }
-    }
-}
-
-impl std::fmt::Display for LandStepKind {
-    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.pad(self.as_str())
-    }
-}
-
-/// Status of a land run or one of its steps.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum LandStatus {
-    Pending,
-    Running,
-    Succeeded,
-    Failed,
-}
-
-impl LandStatus {
-    fn as_str(self) -> &'static str {
-        match self {
-            LandStatus::Pending => "pending",
-            LandStatus::Running => "running",
-            LandStatus::Succeeded => "succeeded",
-            LandStatus::Failed => "failed",
-        }
-    }
-}
-
-impl std::fmt::Display for LandStatus {
-    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.pad(self.as_str())
-    }
-}
-
-pub fn apply_land_from_artifact(artifact_path: &Path, out_path: Option<&Path>) -> Result<()> {
-    let cwd = std::env::current_dir().context("failed to read current directory")?;
-    let mut bundle: crate::model::ChangeGroup = read_json(artifact_path)
-        .with_context(|| format!("failed to load bundle artifact {}", artifact_path.display()))?;
-    if bundle.repos.is_empty() {
-        bail!("Bundle artifact has no repos.");
-    }
-    if bundle.publications.is_empty() {
-        bail!("Bundle artifact has no review publications. Run publish first.");
-    }
-
-    let started_at = now_iso();
-    let mut merged_repo_ids = Vec::new();
-    let mut publication_urls = Vec::new();
-
-    let repos = bundle.repos.clone();
-
-    for repo in &repos {
-        let Some(publication) = publication_for_repo(&bundle, &repo.id).cloned() else {
-            continue;
-        };
-        let forge = providers::for_repo(repo)?;
-        let target = artifact_target(&cwd, forge.as_ref(), repo)?;
-
-        let pr = forge.view(&target, &publication.url)?;
-        if state_is_merged(&pr) {
-            providers::upsert_publication(&mut bundle, repo, forge.as_ref(), &pr);
-            merged_repo_ids.push(repo.id.clone());
-            publication_urls.push(publication.url.clone());
-            println!(
-                "{} {} {}",
-                out::ok("already merged"),
-                out::repo(&repo.id),
-                out::muted(&publication.url)
-            );
-            continue;
-        }
-
-        ensure_open_and_ready(&repo.id, &pr)?;
-
-        let checks_detail = match forge.wait_for_checks(&target, &publication.url, true, 1800, 10) {
-            Ok(summary) => summary.status,
-            Err(err) if providers::is_gh_checks_access_error(&err) => {
-                "passed (checks unavailable)".to_string()
-            }
-            Err(err) => return Err(err),
-        };
-        println!(
-            "{} {} {}",
-            out::ok("checks"),
-            out::repo(&repo.id),
-            out::muted(&checks_detail)
-        );
-
-        forge
-            .merge(
-                &target,
-                &publication.url,
-                MergeMethod::default().as_str(),
-                false,
-                pr.head_ref_oid.as_deref(),
-            )
-            .with_context(|| format!("{}: merging {}", repo.id, publication.url))?;
-
-        let refreshed = forge.view(&target, &publication.url)?;
-        providers::upsert_publication(&mut bundle, repo, forge.as_ref(), &refreshed);
-        merged_repo_ids.push(repo.id.clone());
-        publication_urls.push(publication.url.clone());
-        println!(
-            "{} {} {}",
-            out::ok("merged"),
-            out::repo(&repo.id),
-            out::muted(&publication.url)
-        );
-    }
-
-    // Record a landed node in the artifact without writing land plan/run files.
-    let node = BundleNode::feature_landed(
-        node_id("land"),
-        started_at,
-        format!("land-{}", bundle.id),
-        format!("run-artifact-{}", bundle.id),
-        DEFAULT_LAND_PROVIDER.to_string(),
-        merged_repo_ids,
-        publication_urls,
-    );
-    bundle.nodes.push(node);
-    bundle.head_node_id = bundle.nodes.last().map(|node| node.id.clone());
-    bundle.updated_at = now_iso();
-
-    match out_path {
-        Some(path) => write_json(path, &bundle),
-        None => {
-            let json =
-                serde_json::to_string_pretty(&bundle).context("failed to encode bundle JSON")?;
-            println!("{json}");
-            Ok(())
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct LandPlan {
-    schema_version: String,
-    kind: String,
-    id: String,
-    provider: String,
-    bundle_id: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    source_project_id: Option<String>,
-    created_at: String,
-    steps: Vec<LandStep>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct LandStep {
-    id: String,
-    #[serde(rename = "type")]
-    step_type: LandStepKind,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    needs: Vec<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    repo_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    method: Option<crate::model::MergeMethod>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    wait_for_checks: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    required_checks_only: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    delete_branch: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    required_only: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    timeout_seconds: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    interval_seconds: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    cwd: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    command: Vec<String>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    env: BTreeMap<String, String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    deployment_mode: Option<DeployMode>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    checkout: Option<LandCheckout>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct LandCheckout {
-    branch: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    remote: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    update: Option<DeployCheckoutUpdate>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct LandRun {
-    schema_version: String,
-    kind: String,
-    id: String,
-    plan_id: String,
-    bundle_id: String,
-    provider: String,
-    plan_path: String,
-    status: LandStatus,
-    created_at: String,
-    updated_at: String,
-    steps: Vec<LandRunStep>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct LandRunStep {
-    id: String,
-    #[serde(rename = "type")]
-    step_type: LandStepKind,
-    status: LandStatus,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    repo_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    publication_url: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    started_at: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    finished_at: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    detail: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    stdout: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    stderr: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    exit_code: Option<i32>,
-}
-
-#[derive(Debug)]
-struct StepOutcome {
-    success: bool,
-    detail: String,
-    publication_url: Option<String>,
-    stdout: Option<String>,
-    stderr: Option<String>,
-    exit_code: Option<i32>,
-    publication_update: Option<PublicationUpdate>,
-}
-
-#[derive(Debug)]
-struct PublicationUpdate {
-    repo: RepoEntry,
-    pr: PullRequest,
-}
+use types::*;
 
 pub fn generate_land_plan(
     provider: Option<&str>,
@@ -654,6 +378,7 @@ mod tests {
     use super::validate::ordered_step_ids;
     use super::*;
     use crate::model::SCHEMA_VERSION;
+    use std::collections::BTreeMap;
 
     fn step(id: &str, needs: &[&str]) -> LandStep {
         LandStep {
